DigitalScale/requirements.txt
claid==0.6.4
pydantic~=2.8
opencv-python
scikit-learn
pandas
torch
torchvision
python-dotenv
seaborn
matplotlib
tqdm
python-dotenv
#git+https://github.com/facebookresearch/detectron2.git
wandb


# The following packages are required for the report generation
# jupyter-book~=1.0
# charset-normalizer==2.1.1
# sphinx-inline-tabs
# requests>=2.30
# livereload==2.7.0
# pandas-stubs


# The following packages are required for report v2
mystmd~=1.3
jupyterlab_myst~=2.4
ipywidgets
jupyterlab
matplotlib_venn

commitizen